Update
added verification of the new printer being added before removing the old.

+++ replacePrintServer.ps1
@@ -1,15 +1,20 @@
 
 # print server hostnames
-$oldPrintServer = “OLDSERVERNAME"
-$newPrintServer = “NEWSERVERNAME"
+$oldPrintServer = “OLDPRINTSERVER" # change name to reflect old print server
+$newPrintServer = “NEWPRINTSERVER" # change name to reflect new print server
 
 # get old printers and set new printers including the default printer 
 $oldPrinters = (Get-Printer).Name | where {$_ -match $oldPrintServer} 
 if ($oldPrinters.length -gt 1){
     $newPrinters = $oldPrinters.Replace($oldPrintServer, $newPrintServer)
 }
+
+# get the old default printer and replace old print server name 
+# if the current default printer is in fact on the old server
 $oldDefaultPrinter = (Get-WMIObject -class Win32_Printer -Filter Default=True | Select Name).Name
-$newDefaultPrinter = $oldDefaultPrinter.Replace($oldPrintServer, $newPrintServer)
+if ($oldDefaultPrinter -contains $oldPrintServer){
+    $newDefaultPrinter = $oldDefaultPrinter.Replace($oldPrintServer, $newPrintServer)
+}
 
 # Define functions
 function addPrinters($newPrinters) {
@@ -23,8 +28,17 @@ function addPrinters($newPrinters) {
 function removePrinters($oldPrinters) {
     # remove all old printers 
     foreach ($old in $oldPrinters) {
-        echo "removing $old"
-        Remove-Printer -Name "$old"
+        #verify that the printer is installed before we remove the old printer
+        $tmp = (Get-Printer).Name
+        $new = $old.Replace($oldPrintServer,$newPrintServer)
+        
+        if ($tmp -contains $new){
+            echo "removing $old"
+            Remove-Printer -Name "$old"
+        }
+        else{
+            echo "new printer missing: will not remove $old"
+        }
     }
 }
 
@@ -35,10 +49,17 @@ function setDefaultPrinter($newDefaultPrinter){
 
 # add new and remove the old if there is in fact old printers  
 if ($oldPrinters.Length -gt 1) {
-
+    # add new printers first before anything
     addPrinters($newPrinters)
+    
+    # check to be sure that the new default printer was successfully added
+    # map the new default printer if it exist
+    $tmp = (Get-Printer).Name
+    if ($tmp -contains $oldDefaultPrinter) {
+        setDefaultPrinter($newDefaultPrinter)
+        } 
+    
+    # verify that all the old printers were replaced with new printers before removing the old
     removePrinters($oldPrinters)
-    setDefaultPrinter($newDefaultPrinter)
-    Get-Printer | select name 
 
 }
